requirements.txt
albumentations==0.5.2
easydict==1.11
einops==0.7.0
fastcore==1.5.29
h5py==3.9.0
hydra-core==1.3.2
icecream==2.1.3
imageio==2.31.6
imgaug==0.4.0
invisible_watermark==0.2.0
joblib==1.3.2
kornia==0.7.0
loguru==0.7.2
matplotlib==3.8.0
memory_profiler==0.61.0
nerfacc==0.5.3
numpy==1.24.1
omegaconf==2.3.0
open_clip_torch==2.23.0
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
pandas==2.1.1
Pillow==9.3.0
Pillow==10.1.0
pytorch3d==0.7.4
pytorch_lightning==1.7.7
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.31.0
scikit_learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
setuptools==67.8.0
six==1.16.0
skimage==0.0
sync_batchnorm==0.0.1
timm==0.9.8
tinycudann==1.7
torch==2.0.1+cu118
torch_efficient_distloss==0.1.3
torchvision==0.15.2+cu118
tqdm==4.65.0
transformers==4.34.1
trimesh==4.0.1
webdataset==0.2.62
xformers==0.0.22.post7
